requirements.txt
numpy==1.23.5 
scipy==1.11.4 
py-feat